fix: add method to disable mouse move functionality [DHIS2-6683] (#599)
Implements DHIS2-6683

In the maps app, when dragging the layer cards to sort them, 
the feature labels were being displayed and features were being highlighted
if the card was being dragged over the map.

It should be possible to configure mousemove interactivity on the map. 
So a new method to the Map class has been added to set a flag indicating
whether onMouseMove should execute its actions.